fix lockable_stock
