+++ src/main/java/cam72cam/immersiverailroading/entity/EntityBuildableRollingStock.java
@@ -12,9 +12,9 @@ import cam72cam.immersiverailroading.items.nbt.ItemDefinition;
 import cam72cam.immersiverailroading.items.nbt.ItemGauge;
 import cam72cam.immersiverailroading.items.nbt.ItemPlateType;
 import cam72cam.immersiverailroading.library.AssemblyStep;
+import cam72cam.immersiverailroading.library.ChatText;
 import cam72cam.immersiverailroading.library.ItemComponentType;
 import cam72cam.immersiverailroading.library.StockDeathType;
-import cam72cam.immersiverailroading.library.ChatText;
 import cam72cam.immersiverailroading.net.BuildableStockSyncPacket;
 import cam72cam.immersiverailroading.util.BufferUtil;
 import io.netty.buffer.ByteBuf;
@@ -389,7 +389,7 @@ public class EntityBuildableRollingStock extends EntityRollingStock {
 		if (player.getHeldItem(hand).getItem() == IRItems.ITEM_LARGE_WRENCH || player.getHeldItem(hand).getItem() == IRItems.ITEM_ROLLING_STOCK_COMPONENT) {
 			if (!player.isSneaking()) {
 				addNextComponent(player);
-			} else {
+			} else if (hasPermission(player)) {
 				this.removeNextComponent(player);
 			}
 			return true;
