+++ src/main/java/cam72cam/immersiverailroading/entity/EntityRollingStock.java
@@ -1,119 +1,47 @@
 package cam72cam.immersiverailroading.entity;
 
-import net.minecraftforge.fml.common.FMLCommonHandler;
-import net.minecraftforge.fml.common.network.NetworkRegistry.TargetPoint;
-import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
-import net.minecraftforge.fml.common.registry.IEntityAdditionalSpawnData;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-import net.minecraftforge.fml.server.FMLServerHandler;
-
-import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.UUID;
 
-import com.google.gson.JsonObject;
-
-import cam72cam.immersiverailroading.IRItems;
 import cam72cam.immersiverailroading.ImmersiveRailroading;
-import cam72cam.immersiverailroading.items.ItemLockKey;
 import cam72cam.immersiverailroading.library.Gauge;
-import cam72cam.immersiverailroading.library.LockType;
 import cam72cam.immersiverailroading.library.StockDeathType;
 import cam72cam.immersiverailroading.registry.DefinitionManager;
 import cam72cam.immersiverailroading.registry.EntityRollingStockDefinition;
 import cam72cam.immersiverailroading.util.BufferUtil;
 import io.netty.buffer.ByteBuf;
-import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Items;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.network.datasync.DataParameter;
-import net.minecraft.network.datasync.DataSerializer;
-import net.minecraft.network.datasync.DataSerializers;
-import net.minecraft.network.datasync.EntityDataManager;
-import net.minecraft.server.management.UserListOpsEntry;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.EnumHand;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.common.network.NetworkRegistry.TargetPoint;
+import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
+import net.minecraftforge.fml.common.registry.IEntityAdditionalSpawnData;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
-public abstract class EntityRollingStock extends Entity implements IEntityAdditionalSpawnData {
-	
-	private static DataParameter<String> LOCK_TYPE = EntityDataManager.createKey(EntityRollingStock.class, DataSerializers.STRING);	//0=everyone can enter/break, 1=everyone can enter, 2=no one can enter/break
-	private static DataParameter<String> LOCK_OWNER = EntityDataManager.createKey(EntityRollingStock.class, DataSerializers.STRING);
-	
-	protected String defID;
+public abstract class EntityRollingStock extends EntityLockableRollingStock implements IEntityAdditionalSpawnData {	
 	public Gauge gauge;
 	public String tag = "";
 
 	public EntityRollingStock(World world, String defID) {
-		super(world);
-
-		this.defID = defID;
+		super(world, defID);
 
 		super.preventEntitySpawning = true;
 		super.isImmuneToFire = true;
 		super.entityCollisionReduction = 1F;
 		super.ignoreFrustumCheck = true;
-		
-		this.getDataManager().register(LOCK_TYPE, "UNLOCKED");
-		this.getDataManager().register(LOCK_OWNER, "");
 	}
 	
 	@Override
 	public String getName() {
 		return this.getDefinition().name();
 	}
-
-	public EntityRollingStockDefinition getDefinition() {
-		return this.getDefinition(EntityRollingStockDefinition.class);
-	}
-	public <T extends EntityRollingStockDefinition> T getDefinition(Class<T> type) {
-		EntityRollingStockDefinition def = DefinitionManager.getDefinition(defID);
-		if (def == null) {
-			try {
-				return type.getConstructor(String.class, JsonObject.class).newInstance(defID, (JsonObject)null);
-			} catch (InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException | NoSuchMethodException
-					| SecurityException e) {
-				e.printStackTrace();
-				return null;
-			}
-		} else {
-			return type.cast(def);
-		}
-	}
+	
 	public String getDefinitionID() {
 		return this.defID;
-	}
-	
-	public LockType getLockType () {
-		return LockType.valueOf(this.getDataManager().get(LOCK_TYPE));
-	}
-	
-	public void setLockType (LockType type) {
-		this.getDataManager().set(LOCK_TYPE, type.toString());
-	}
-	
-	public UUID getLockOwner() {
-		if (this.getDataManager().get(LOCK_OWNER) == "") {
-			return null;
-		}
-		return UUID.fromString(this.getDataManager().get(LOCK_OWNER));
-	}
-	
-	public void setLockOwner(String playerId) {
-		this.getDataManager().set(LOCK_OWNER, playerId);
-	}
-	
-	public void switchLockType (EntityPlayer player) {
-		setLockType(getLockType().next());
-		setLockOwner(EntityPlayer.getUUID(player.getGameProfile()).toString());
 	}
 	
 	@Override
@@ -124,14 +52,6 @@ public abstract class EntityRollingStock extends Entity implements IEntityAdditi
 				world.removeEntity(this);
 			}
 		}
-	}
-	
-	public boolean playerHasOp (EntityPlayer player) {
-		return  FMLCommonHandler.instance().getMinecraftServerInstance().getPlayerList().getOppedPlayers().getEntry(player.getGameProfile()) != null;
-	}
-	
-	public boolean hasPermission (EntityPlayer player) {
-		return getLockType() == LockType.UNLOCKED || (getLockOwner().equals(player.getUniqueID()) || getLockOwner() == null) || playerHasOp(player);
 	}
 
 	/*
@@ -155,15 +75,15 @@ public abstract class EntityRollingStock extends Entity implements IEntityAdditi
 
 	@Override
 	protected void writeEntityToNBT(NBTTagCompound nbt) {
+		super.writeEntityToNBT(nbt);
 		nbt.setString("defID", defID);
 		nbt.setDouble("gauge", gauge.value());
 		nbt.setString("tag", tag);
-		nbt.setString("lock_type", getLockType().toString());
-		nbt.setString("lock_owner", getLockOwner() != null ? getLockOwner().toString() : "");
 	}
 
 	@Override
 	protected void readEntityFromNBT(NBTTagCompound nbt) {
+		super.readEntityFromNBT(nbt);
 		defID = nbt.getString("defID");
 		if (nbt.hasKey("gauge")) {
 			gauge = Gauge.from(nbt.getDouble("gauge"));
@@ -172,8 +92,6 @@ public abstract class EntityRollingStock extends Entity implements IEntityAdditi
 		}
 		
 		tag = nbt.getString("tag");
-		setLockType(LockType.valueOf(nbt.getString("lock_type")));
-		setLockOwner(nbt.getString("lock_owner"));
 	}
 
 	@Override
@@ -186,11 +104,7 @@ public abstract class EntityRollingStock extends Entity implements IEntityAdditi
 	
 	@Override
 	public boolean processInitialInteract(EntityPlayer player, EnumHand hand) {
-		if (player.getHeldItem(hand).getItem() == IRItems.ITEM_LOCK_KEY && hasPermission(player)) {
-			switchLockType(player);
-			return true;
-		}
-		return false;
+		return super.processInitialInteract(player, hand);
 	}
 
 	@Override
@@ -221,7 +135,7 @@ public abstract class EntityRollingStock extends Entity implements IEntityAdditi
 		
 		if (damagesource.getTrueSource() instanceof EntityPlayer && !damagesource.isProjectile()) {
 			EntityPlayer player = (EntityPlayer) damagesource.getTrueSource();
-			if (player.isSneaking() && hasPermission(player)) {
+			if (player.isSneaking() && canRide(player)) {
 				if (!this.isDead) {
 					this.onDeath(StockDeathType.PLAYER);
 				}
